+++ js-miniapp-sdk/test/miniapp.spec.ts
@@ -11,24 +11,30 @@ import {
 } from '../../js-miniapp-bridge/src';
 import { MiniApp } from '../src/miniapp';
 
+const sandbox = sinon.createSandbox();
+beforeEach(() => {
+  sandbox.restore();
+  sandbox.reset();
+});
+
 const window: any = {};
 (global as any).window = window;
 
 window.MiniAppBridge = {
-  getUniqueId: sinon.stub(),
-  requestPermission: sinon.stub(),
-  requestCustomPermissions: sinon.stub(),
-  loadInterstitialAd: sinon.stub(),
-  loadRewardedAd: sinon.stub(),
-  showInterstitialAd: sinon.stub(),
-  showRewardedAd: sinon.stub(),
-  shareInfo: sinon.stub(),
-  getPlatform: sinon.stub(),
-  getUserName: sinon.stub(),
-  getProfilePhoto: sinon.stub(),
-  getContacts: sinon.stub(),
-  getAccessToken: sinon.stub(),
-  setScreenOrientation: sinon.stub(),
+  getUniqueId: sandbox.stub(),
+  requestPermission: sandbox.stub(),
+  requestCustomPermissions: sandbox.stub(),
+  loadInterstitialAd: sandbox.stub(),
+  loadRewardedAd: sandbox.stub(),
+  showInterstitialAd: sandbox.stub(),
+  showRewardedAd: sandbox.stub(),
+  shareInfo: sandbox.stub(),
+  getPlatform: sandbox.stub(),
+  getUserName: sandbox.stub(),
+  getProfilePhoto: sandbox.stub(),
+  getContacts: sandbox.stub(),
+  getAccessToken: sandbox.stub(),
+  setScreenOrientation: sandbox.stub(),
 };
 const miniApp = new MiniApp();
 
@@ -42,17 +48,20 @@ describe('getUniqueId', () => {
   });
 });
 
-describe('requestPermission', () => {
-  window.MiniAppBridge.requestCustomPermissions.resolves({
-    permissions: [
-      {
-        name: CustomPermissionName.LOCATION,
-        status: CustomPermissionStatus.ALLOWED,
-      },
-    ],
+describe('requestLocationPermission', () => {
+  beforeEach(() => {
+    window.MiniAppBridge.requestCustomPermissions.resolves({
+      permissions: [
+        {
+          name: CustomPermissionName.LOCATION,
+          status: CustomPermissionStatus.ALLOWED,
+        },
+      ],
+    });
+    window.MiniAppBridge.requestPermission.resolves('Accept');
   });
 
-  it('should delegate to requestPermission function when request any permission', () => {
+  it('should delegate to requestPermission function when request location permission', () => {
     const spy = sinon.spy(miniApp, 'requestPermission' as any);
 
     return miniApp
@@ -65,6 +74,43 @@ describe('requestPermission', () => {
 
     return expect(miniApp.requestLocationPermission()).to.eventually.equal(
       'Denied'
+    );
+  });
+
+  it('should request location custom permission', () => {
+    return miniApp.requestLocationPermission('test_description')
+      .then(() => {
+        sinon.assert.calledWith(window.MiniAppBridge.requestCustomPermissions, [
+          {
+            name: CustomPermissionName.LOCATION,
+            description: 'test_description',
+          },
+        ]);
+      })
+  });
+
+  it('should handle case where Android SDK does not support location custom permission', () => {
+    window.MiniAppBridge.requestCustomPermissions.resolves({
+      permissions: [
+        {
+          name: CustomPermissionName.LOCATION,
+          status: CustomPermissionStatus.PERMISSION_NOT_AVAILABLE,
+        },
+      ],
+    });
+
+    return expect(miniApp.requestLocationPermission()).to.eventually.equal(
+      'Accept'
+    );
+  });
+
+  it('should handle case where iOS SDK does not support location custom permission', () => {
+    window.MiniAppBridge.requestCustomPermissions.returns(
+      Promise.reject('invalidCustomPermissionsList: test description')
+    );
+
+    return expect(miniApp.requestLocationPermission()).to.eventually.equal(
+      'Accept'
     );
   });
 });